fix compile error

+++ handler/httphandler/CGI.cpp
@@ -154,7 +154,7 @@ int CGI::write()
 		entrustCreateResponse(INTERNAL_SERVER_ERROR);
 		return -1;
 	}
-	if (nb < _buff.size()) {
+	if (static_cast<size_t>(nb) < _buff.size()) {
 		_buff = _buff.substr(0, nb);
 		return 1;
 	}
